Also implement the fix for multiline `__slots__` and `__match_args__`
